backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg[binary]>=3.3.2
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
drf-yasg==1.21.7
django-model-utils==4.4.0
firebase-admin>=6.0.0
PyJWT>=2.8.0